refactor(authz): rename the seam to ./authz and nest the OIDC adapter
The seam is method-agnostic — the AuthorizationServerAdapter port, the
registry, and the AuthorizationRequirement union (oauth/mdl/idv/credential/
none) are not OAuth-specific. Rename the ./oauth entry point to ./authz so the
name reflects the abstraction, and nest the OAuth/OIDC concrete adapter under
authz/oidc/ so future credential/mdl/idv adapters are siblings. Pure rename;
the export name has not shipped, so no consumer is affected.
