+++ src/authz/registry.ts
